Fix the create_message function for clients with no default service.
Even though the Client.create_message function expects a service
parameter, this service was not passed on to the SOAP implementation.
The SOAP binding would invariably use the default service in the client
object, and raise an exception if it was not found.

Fix this by adding an optional service parameter to the
SoapBinding.create method as well.

Fixes #833