signature: Allow specifying a the signing key used by the SOAP server
Before this change, zeep assumed that the SMIME signing key used by
client and server was the same. This is generally false, and made it
basically impossible to implement clients that use a private signing key
to exchange signed messages with a public SOAP server that uses its own
signing key (which is never shared with clients).